move the test to 'envoyconfig' dir so it's easier to extend for other policies
